Remove unnecessary intval() and add missing (int) casts
intval() is about 5 times slower than a type cast and should be avoided
if not necessary. Comparisons with == do not need an explicit type cast,
this happens automatically.

This patch also adds some (int) type casts in places where it can turn
into a security issue if a string is used as part of a WHERE clause.

+++ lib/includes/PidLock.php
@@ -87,7 +87,7 @@ class PidLock {
 			// check if the process still exist and is alive
 			// XXX: there's a race condition here.
 			if ( file_exists( $file ) ) {
-				$pid = file_get_contents( $file );
+				$pid = (int)file_get_contents( $file );
 				if ( $this->isAlive( $pid ) === true ) {
 					return false;
 				}
